🤖 Refactor for ESLint
