+++ index.js
@@ -1,8 +1,8 @@
 #!/usr/bin/env node
 
+import FindActionUses from './utils/FindActionUses.mjs'
 import chalk from 'chalk'
 import meow from 'meow'
-import FindActionUses from './utils/FindActionUses.js'
 
 const {dim, blue, bold, red, yellow} = chalk
 const cli = meow(
