+++ prettier.config.js
@@ -1,1 +1,2 @@
+// eslint-disable-next-line import/extensions, import/no-commonjs
 module.exports = require('eslint-plugin-github/prettier.config')
